PS4 Controller Pairing
This implements the ability to pair with the PS4 controller and interpret the inputs from the controller. The inputs have not yet been associated with specific motor control functions, so this is for demonstration alone. Later I will clean up the demonstration functions and streamline the program as a whole, associating specific controller inputs with the perspective motor control functions.