Added PS4 Controller Battery Indication
The PS4 controller's battery level is evaluated on a scale of 0-15. 
If the value returned is greater than 8, the PS4 LED will be set to green.
If the value returned is between 4 and 8, the PS4 LED will be set to yellow.
If the value returned is less than 4, the PS4 LED will be set to red.

This requires further testing but appears to be functioning as intended.


Also trimmed out a variety of PS4-based information that had already been commented out.