Controller Failsafe Update
Updated BTHID.cpp and BTHID.h library files to implement a feature that will allow a failsafe in the event of unintended controller connection loss. If the controller does not send any information in 1000ms, the failsafe will activate, return control to the child, and set the rumble on the ps4 controller to high.