Priority Control Updates and Etc.
A priority control system has been added such that when wireless is enabled, it will take full priority. If no wireless inputs are received for a set period (currently 5s) then the control is returned to the child. Upon entering a wireless input, control is immediately returned to the wireless controller.

Other various additions such as a beta version of a range detection/warning system using the controller rumble.